refactor to have ReleaseNoteFormatter in utilities
